Refactor outlier angles detection function
The outlier detection has been slimmed down using apply functions, and
altered to return a boolean vector instead of mutating the input data-frame.

A duplicate instance of the function has been removed.

The NAMESPACE has been updated to export all functions starting with a letter
when the package is loaded.